refactor: switch to a monorepo and create individual packages for each adapter
