+++ .github/workflows/test.yml
@@ -8,5 +8,6 @@ jobs:
         uses: actions/checkout@master
       - name: Docker
         run: npm run test:docker
-      - name: Posttest
-        run: npm run posttest:docker
+      - name: Coverage
+        if: success()
+        run: npm run coverage
