+++ Dockerfile
@@ -1,4 +1,5 @@
 FROM node:lts-alpine
 WORKDIR /repo
 CMD npm install && \
-    npm test
+    npm test && \
+    npm run coverage:html
