+++ package.json
@@ -1,33 +1,18 @@
 {
-  "name": "endb",
+  "private": true,
+  "name": "@endbjs/monorepo",
   "version": "0.22.9",
   "description": "Simple key-value storage with support for multiple backends.",
-  "main": "src/index.js",
-  "typings": "typings/index.d.ts",
-  "files": [
-    "src",
-    "typings",
-    "esm"
-  ],
-  "exports": {
-    ".": [
-      {
-        "require": "./src/index.js",
-        "import": "./esm/index.mjs"
-      },
-      "./src/index.js"
-    ],
-    "./esm": "./esm/index.mjs"
-  },
   "scripts": {
     "lint": "xo src",
     "lint:fix": "xo src --fix",
     "test": "nyc ava --serial",
     "coverage": "nyc report --reporter=text-lcov | coveralls",
+    "coverage:html": "nyc report --reporter=html",
     "test:docker": "docker-compose up --build --abort-on-container-exit --exit-code-from test",
     "posttest:docker": "docker-compose down --rmi local",
-    "docs:build": "node docs.js",
-    "prepublishOnly": "npm run lint && nyc ava test/index.js"
+    "prepublishOnly": "npm run lint && nyc ava test/index.js",
+    "release": "lerna publish"
   },
   "keywords": [
     "endb",
@@ -40,21 +25,12 @@
   "homepage": "https://endb.js.org",
   "repository": "chroventer/endb",
   "bugs": "https://github.com/chroventer/endb/issues",
-  "dependencies": {
-    "buffer-json": "^2.0.0",
-    "lodash": "^4.17.15"
-  },
+  "dependencies": {},
   "devDependencies": {
     "ava": "^3.7.1",
     "coveralls": "^3.1.0",
-    "docma": "^3.2.2",
-    "mongodb": "^3.5.6",
-    "mysql2": "^2.1.0",
+    "lerna": "^3.20.2",
     "nyc": "^15.0.1",
-    "pg": "^8.0.2",
-    "redis": "^3.0.2",
-    "sql": "^0.78.0",
-    "sqlite3": "^4.1.1",
     "xo": "^0.30.0"
   },
   "ava": {
