+++ packages/endb-mongodb/README.md
@@ -1,0 +1,11 @@
+# `endb-mongodb`
+
+> TODO: description
+
+## Usage
+
+```
+const endbMongodb = require('endb-mongodb');
+
+// TODO: DEMONSTRATE API
+```
