+++ packages/endb-mongodb/__tests__/endb-mongodb.test.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const endbMongodb = require('..');
+
+describe('endb-mongodb', () => {
+    it('needs tests');
+});
