+++ packages/endb-mysql/README.md
@@ -1,0 +1,11 @@
+# `endb-mysql`
+
+> TODO: description
+
+## Usage
+
+```
+const endbMysql = require('endb-mysql');
+
+// TODO: DEMONSTRATE API
+```
