+++ packages/endb-mysql/__tests__/endb-mysql.test.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const endbMysql = require('..');
+
+describe('endb-mysql', () => {
+    it('needs tests');
+});
