+++ packages/endb-postgres/__tests__/endb-postgres.test.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const endbPostgres = require('..');
+
+describe('endb-postgres', () => {
+    it('needs tests');
+});
