+++ packages/endb-redis/__tests__/endb-redis.test.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const endbRedis = require('..');
+
+describe('endb-redis', () => {
+    it('needs tests');
+});
