+++ packages/endb-redis/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "endb-redis",
+  "version": "1.0.0",
+  "description": "> TODO: description",
+  "author": "chroventer <ayyanlewis@gmail.com>",
+  "homepage": "https://endb.js.org/packages/endb-redis",
+  "license": "ISC",
+  "main": "lib/endb-redis.js",
+  "directories": {
+    "lib": "lib",
+    "test": "__tests__"
+  },
+  "files": [
+    "lib"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/chroventer/endb.git"
+  },
+  "scripts": {
+    "test": "echo \"Error: run tests from root\" && exit 1"
+  },
+  "bugs": {
+    "url": "https://github.com/chroventer/endb/issues"
+  }
+}
