+++ packages/endb-sql/README.md
@@ -1,0 +1,11 @@
+# `endb-sql`
+
+> TODO: description
+
+## Usage
+
+```
+const endbSql = require('endb-sql');
+
+// TODO: DEMONSTRATE API
+```
