+++ packages/endb-sql/__tests__/endb-sql.test.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const endbSql = require('..');
+
+describe('endb-sql', () => {
+    it('needs tests');
+});
