+++ packages/endb-sql/package.json
@@ -1,0 +1,26 @@
+{
+  "name": "endb-sql",
+  "version": "1.0.0",
+  "description": "Parent class for SQL-based adapter for Endb.",
+  "main": "src/index.js",
+  "files": [
+    "src"
+  ],
+  "scripts": {
+    "test": "echo \"Error: run tests from root\" && exit 1"
+  },
+  "author": "chroventer (https://github.com/chroventer)",
+  "license": "MIT",
+  "repository": "github:chroventer/endb",
+  "homepage": "https://github.com/endbjs/endb#homepage",
+  "bugs": "https://github.com/endbjs/endb/issues",
+  "keywords": [
+    "endb",
+    "sql",
+    "adapter",
+    "sqlite",
+    "postgresql",
+    "mysql",
+    "key-value"
+  ]
+}
