+++ packages/endb-sqlite/README.md
@@ -1,0 +1,11 @@
+# `endb-sqlite`
+
+> TODO: description
+
+## Usage
+
+```
+const endbSqlite = require('endb-sqlite');
+
+// TODO: DEMONSTRATE API
+```
