+++ packages/endb-sqlite/__tests__/endb-sqlite.test.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const endbSqlite = require('..');
+
+describe('endb-sqlite', () => {
+    it('needs tests');
+});
