+++ packages/endb/__tests__/endb.test.js
@@ -1,0 +1,7 @@
+'use strict';
+
+const endb = require('..');
+
+describe('endb', () => {
+    it('needs tests');
+});
